perf(serve): store GDN recurrent state in bf16, compute unchanged
The recurrent state was the largest single cost in a decode round — 2 MiB per
lane per layer, ~6 GiB moved per round at 64 lanes, 22.6% of device time at 71%
of peak bandwidth — and it was stored fp32 while vLLM stores it at the model
dtype (gated_delta_net_state_dtype resolves 'auto' to bf16). Compute still runs
fp32 in registers; only the HBM representation narrows, at the load/store seam.
GdnStateStorage names the choice so it can be reverted in one line.

4B 2,666 -> 3,024 tok/s (+13%), 0.8B 6,047 -> 6,646 (+10%), zero errors, output
verified coherent at temperature 0. PATCHES.md #57.

+++ csrc/src/serve/core/linear_attention_state.cpp
@@ -95,7 +95,7 @@ plan_linear_attention_state_pool(LayoutBuilder& builder, const LinearAttentionSt
     const Tensor conv_shape(nullptr, spec.conv_dtype,
                             {spec.conv_channels, spec.conv_width, spec.slot_count});
     const Tensor recurrent_shape(
-        nullptr, DType::FP32,
+        nullptr, DType::BF16,
         {spec.key_head_dim, spec.value_head_dim, spec.value_heads, spec.slot_count});
 
     LinearAttentionStatePoolLayout layout;
@@ -123,7 +123,7 @@ LinearAttentionStatePool::LinearAttentionStatePool(DeviceSpan backing,
     const Tensor conv_shape(nullptr, spec.conv_dtype,
                             {spec.conv_channels, spec.conv_width, spec.slot_count});
     const Tensor recurrent_shape(
-        nullptr, DType::FP32,
+        nullptr, DType::BF16,
         {spec.key_head_dim, spec.value_head_dim, spec.value_heads, spec.slot_count});
     conv.reserve(layout.conv.size());
     recurrent.reserve(layout.recurrent.size());
@@ -137,7 +137,7 @@ LinearAttentionStatePool::LinearAttentionStatePool(DeviceSpan backing,
                           std::initializer_list<std::int32_t>{spec.conv_channels, spec.conv_width,
                                                               spec.slot_count});
         recurrent.emplace_back(
-            layout.recurrent[layer].bind(backing).data, DType::FP32,
+            layout.recurrent[layer].bind(backing).data, DType::BF16,
             std::initializer_list<std::int32_t>{spec.key_head_dim, spec.value_head_dim,
                                                 spec.value_heads, spec.slot_count});
     }
@@ -168,7 +168,7 @@ LinearAttentionStateAllLayersView LinearAttentionStatePool::all_layers_view() co
         validate_state_tensor(conv[layer], spec.conv_dtype,
                               {spec.conv_channels, spec.conv_width, spec.slot_count}, "conv");
         validate_state_tensor(
-            recurrent[layer], DType::FP32,
+            recurrent[layer], DType::BF16,
             {spec.key_head_dim, spec.value_head_dim, spec.value_heads, spec.slot_count},
             "recurrent");
     }
