+++ csrc/src/serve/ops/linear_attention/gated_delta_net/chunked/state_passing.cuh
@@ -172,9 +172,9 @@ __launch_bounds__(kernel_dims<NStrip>::THREADS, kernel_dims<NStrip>::MIN_BLOCKS)
     void state_passing_kernel(const __nv_bfloat16* __restrict__ W_in,
                               const __nv_bfloat16* __restrict__ U_in,
                               const __nv_bfloat16* __restrict__ k_in,
-                              const float* __restrict__ g_cumsum, const float* state_in,
+                              const float* __restrict__ g_cumsum, const __nv_bfloat16* state_in,
                               __nv_bfloat16* __restrict__ v_new,
-                              __nv_bfloat16* __restrict__ h_chunk, float* state_out,
+                              __nv_bfloat16* __restrict__ h_chunk, __nv_bfloat16* state_out,
                               head_map qk_map, int chunks) {
     using D                         = kernel_dims<NStrip>;
     using L                         = smem_layout<NStrip>;
@@ -263,13 +263,13 @@ __launch_bounds__(kernel_dims<NStrip>::THREADS, kernel_dims<NStrip>::MIN_BLOCKS)
             const int col_d0 = warp_d_global + 2 * lane_t;
             const int col_d1 = col_d0 + 1;
             h_frag[m][0] =
-                load_ldg<float>(state_in + st_base + (int64_t)col_d0 * kStateDim + row_g0);
+                __bfloat162float(load_ldg<__nv_bfloat16>(state_in + st_base + (int64_t)col_d0 * kStateDim + row_g0));
             h_frag[m][1] =
-                load_ldg<float>(state_in + st_base + (int64_t)col_d1 * kStateDim + row_g0);
+                __bfloat162float(load_ldg<__nv_bfloat16>(state_in + st_base + (int64_t)col_d1 * kStateDim + row_g0));
             h_frag[m][2] =
-                load_ldg<float>(state_in + st_base + (int64_t)col_d0 * kStateDim + row_g1);
+                __bfloat162float(load_ldg<__nv_bfloat16>(state_in + st_base + (int64_t)col_d0 * kStateDim + row_g1));
             h_frag[m][3] =
-                load_ldg<float>(state_in + st_base + (int64_t)col_d1 * kStateDim + row_g1);
+                __bfloat162float(load_ldg<__nv_bfloat16>(state_in + st_base + (int64_t)col_d1 * kStateDim + row_g1));
         }
     }
 
@@ -528,10 +528,10 @@ __launch_bounds__(kernel_dims<NStrip>::THREADS, kernel_dims<NStrip>::MIN_BLOCKS)
         const int k_g1 = k_g0 + 8;
         const int d0   = warp_d_global + 2 * lane_t;
         const int d1   = d0 + 1;
-        state_out[st_base + (int64_t)d0 * kStateDim + k_g0] = h_frag[m][0];
-        state_out[st_base + (int64_t)d1 * kStateDim + k_g0] = h_frag[m][1];
-        state_out[st_base + (int64_t)d0 * kStateDim + k_g1] = h_frag[m][2];
-        state_out[st_base + (int64_t)d1 * kStateDim + k_g1] = h_frag[m][3];
+        state_out[st_base + (int64_t)d0 * kStateDim + k_g0] = __float2bfloat16_rn(h_frag[m][0]);
+        state_out[st_base + (int64_t)d1 * kStateDim + k_g0] = __float2bfloat16_rn(h_frag[m][1]);
+        state_out[st_base + (int64_t)d0 * kStateDim + k_g1] = __float2bfloat16_rn(h_frag[m][2]);
+        state_out[st_base + (int64_t)d1 * kStateDim + k_g1] = __float2bfloat16_rn(h_frag[m][3]);
     }
 }
